// src/app/model/tokenRequest.ts
/**
 * Customer accounts API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: @springdoc.version@
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { RequestedClaim } from './requestedClaim';


export interface TokenRequest { 
    subject?: string;
    tokenName?: string;
    audience?: string;
    notBefore?: Date;
    expiration?: Date;
    issuer?: string;
    claims?: Array<RequestedClaim>;
    key?: string;
    fromKeystore?: boolean;

}