// src/app/model/appClaim.ts
/**
 * Customer accounts API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: @springdoc.version@
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 export interface AppClaim { 
    name?: string;
    value?: string;
}